# docs/swagger.yaml
basePath: /
definitions:
  model.Tag:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: tag name
        type: string
    type: object
  tagrequest.CreateTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  tagresponse.FindAllTagResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      page:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: Minh Thuong
  description: This is a sample server Cake ATS API server.
  title: Cake ATS API
  version: "1.0"
paths:
  /tags:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all tags with pagination support
      parameters:
      - description: Page number (default is 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default is 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of tags
          schema:
            $ref: '#/definitions/tagresponse.FindAllTagResponse'
      summary: Retrieve all tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: Create a new tag based on the provided JSON data
      parameters:
      - description: Tag data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/tagrequest.CreateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tag created successfully
          schema:
            $ref: '#/definitions/model.Tag'
      summary: Create a new tag
      tags:
      - tags
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"